archive_code/naive_agent/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# naive_agent/data_preprocess.py: 12,13
# naive_agent/imitation_play.py: 11
# naive_agent/imitation_train.py: 9
matplotlib == 3.3.3

# naive_agent/data_preprocess.py: 8
# naive_agent/gen_play_data.py: 9
# naive_agent/imitation_play.py: 9
numpy == 1.19.5

# naive_agent/imitation_model.py: 10
# naive_agent/imitation_play.py: 10
# naive_agent/imitation_train.py: 8
# tensorflow == 2.4.1
# or, with gpu, you know :p
tensorflow_gpu == 2.4.1

# naive_agent/basic.py: 15
# naive_agent/gen_play_data.py: 8
# naive_agent/imitation_play.py: 8
# naive_agent/spec_test.py: 15
vizdoom == 1.1.8
